+++ src/Config/PostgresDriverConfig.php
@@ -49,6 +49,7 @@ class PostgresDriverConfig extends DriverConfig
         bool $queryCache = true,
         bool $readonlySchema = false,
         bool $readonly = false,
+        bool $datetimeWithMicroseconds = false,
     ) {
         /** @psalm-suppress ArgumentTypeCoercion */
         parent::__construct(
@@ -59,6 +60,7 @@ class PostgresDriverConfig extends DriverConfig
             queryCache: $queryCache,
             readonlySchema: $readonlySchema,
             readonly: $readonly,
+            datetimeWithMicroseconds: $datetimeWithMicroseconds,
         );
 
         $this->schema = $this->bootSchema($schema);