+++ tests/Database/Functional/Driver/Common/Schema/DefaultValueTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Cycle\Database\Tests\Functional\Driver\Common\Schema;
 
-use Cycle\Database\Schema\AbstractTable;
 use Cycle\Database\Tests\Functional\Driver\Common\BaseTest;
 
 /**
@@ -17,11 +16,6 @@ use Cycle\Database\Tests\Functional\Driver\Common\BaseTest;
  */
 abstract class DefaultValueTest extends BaseTest
 {
-    public function schema(string $table): AbstractTable
-    {
-        return $this->database->table($table)->getSchema();
-    }
-
     public function testDefaultNullValueForInteger(): void
     {
         $schema = $this->schema('table');